requirements.txt
numpy
imutils
opencv-python
sklearn.neighbours